all issues fixed, pattern strategy added v1.3
