+++ ARCHITECTURE.md
@@ -9,6 +9,7 @@
     4. Класс Game реализует очередность ходов и считывание данных
     
     5. Для удобного спауна новых юнитов существует класс-адаптер AdaptedBarracks, его название говорит само за себя 
+    6. Для возможности удобной смены алгоритма поиска расстояния использован паттерн стратегия (pathfinding)
 
 ## Используемые паттерны
     - Порождающие:
@@ -17,6 +18,7 @@
         2. Адаптер
     - Поведенческие:
         3. Command
+        4. strategy
 
 ## причины отказа от других паттернов:
