+++ src/buildings/CMakeLists.txt
@@ -1,6 +1,5 @@
 cmake_minimum_required(VERSION 3.1)
 project(buildings)
 
-set(SOURCE_LIB barracks.cpp)
+set(SRC ${SRC} adapted_barracks.cpp adapted_barracks.h barracks.h barracks.cpp)
 
-add_library(buildings STATIC ${SOURCE_LIB} adapted_barracks.cpp)
