+++ src/buildings/barracks
@@ -1,3 +1,3 @@
 #pragma once
-#include "barracks.cpp"
-#include "adapted_barracks.cpp"
+#include "barracks.h"
+#include "adapted_barracks.h"
