+++ src/commands/command_attack.h
@@ -1,3 +1,4 @@
+#pragma once
 #include "command.h"
 
 class CommandAttack : public Command {
