+++ src/commands/commandbarracks.h
@@ -1,3 +1,4 @@
+#pragma once
 #include "command.h"
 
 class CommandBarracks : public Command {
