+++ src/commands/commands
@@ -1,4 +1,4 @@
 #pragma once
-#include "command_attack.cpp"
-#include "commandbarracks.cpp"
-#include "moveunit.cpp" 
+#include "command_attack.h"
+#include "commandbarracks.h"
+#include "moveunit.h"
