+++ src/map/CMakeLists.txt
@@ -1,5 +1,4 @@
 cmake_minimum_required(VERSION 3.1)
 project(gamemap)
 
-set(SOURCE_LIB gamemap.cpp)
-add_library(gamemap STATIC ${SOURCE_LIB}) 
+set(SRC ${SRC} gamemap.cpp map.h map.cpp gamemap.h)
