--- src/map/gamemap
@@ -1,2 +1,0 @@
-#pragma once
-#include "gamemap.cpp"
